+++ flow-typed/interfaces/engine-capabilty.js
@@ -1,0 +1,7 @@
+//@flow
+declare type CapabilityResult = { [capabilityName: string]: any };
+
+declare interface ICapability {
+  static runCapability(): void;
+  static getCapability(): Promise<CapabilityResult>;
+}